Server/requirements.txt
fastapi
uvicorn[standard]
motor
pydantic
python-dotenv
passlib[bcrypt]
httpx
tensorflow
opencv-python
pandas
numpy
python-multipart

